pyproject.toml
[project]
name = "tui-assistant"
version = "1.0.0"
description = "私人语音助理'小跟班' - 树莓派智能助手"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # 音频处理
    "pyaudio>=0.2.14",
    "numpy>=1.24.0",

    # 唤醒词检测
    "openwakeword>=0.5.0",

    # VAD语音活动检测
    "torch>=2.0.0",
    "torchaudio>=2.0.0",

    # 腾讯云SDK
    "tencentcloud-sdk-python>=3.0.0",

    # OpenAI协议客户端
    "openai>=1.12.0",

    # TTS
    "edge-tts>=6.1.0",

    # 配置管理
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # 日志
    "loguru>=0.7.2",

    # 异步工具
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
